Framework for a NatBC executable
The BC means ByteCode, which is once again open for improvements. This
first version of the executable is a simple hello world program that is
created by transpiling the example script to CPP and removing the parts
we don't need. This means there is a lot of duplication with
`src/main.cpp`.