Rewrite StringObject::append with StringView to copy instead of reference
References are sneakily converted to pointers under the hood. In the case of a
StringView, it's more efficient to just copy the tiny data structure than to add
a layer of indirection by passing a reference.

A quick test with `p "foo12".swapcase` (the `String#swapcase` method is
one of the few direct callers of this append function) removed roughly
6000 instructions in callgrind. A big part of that would probably be
internal use.