Don't convert StringView to String in StringObject::append
All the data required to append is already available via the string view itself,
there is no need to copy it into a new string object just to throw that object
away after the copy. The string uses heap allocation for the chars, so this
remove a new+delete.
This removes another 300 instructions from our previous test case.